moved the location of tes numpy utils
